+++ src/main/java/com/example/springRestAWS/resource/UserResource.java
@@ -1,14 +1,18 @@
 package com.example.springRestAWS.resource;
 
 import com.example.springRestAWS.exception.RestrictedInfoException;
+import com.example.springRestAWS.model.FirebaseUser;
 import com.example.springRestAWS.model.User;
+import com.example.springRestAWS.service.FirebaseService;
 import com.example.springRestAWS.service.UserService;
+import com.google.firebase.auth.FirebaseAuthException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.io.IOException;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -20,10 +24,18 @@ public class UserResource {
     @Autowired
     UserService userService;
 
+    @Autowired
+    FirebaseService firebaseService;
+
     //POST API /CREATE
     @PostMapping
-    public User saveUser(@RequestBody @Valid User user){
-        return userService.saveUser(user);
+    public User saveUser(@RequestBody User user,@RequestHeader(name="idToken")String idToken) throws IOException, FirebaseAuthException {
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+        if(firebaseUser != null) {
+            return userService.saveUser(user);
+        }else{
+            return null;
+        }
     }
 
     //GET API / View
@@ -59,14 +71,22 @@ public class UserResource {
 
     //PUT API / update
     @PutMapping
-    public User updateUser(@RequestBody User user){
-        return userService.updateUser(user);
+    public User updateUser(@RequestBody User user,@RequestHeader(name="idToken")String idToken) throws IOException, FirebaseAuthException {
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+        if (firebaseUser != null){
+            return userService.updateUser(user);
+        }else{
+            return null;
+        }
     }
 
     //DELETE API
     @DeleteMapping
-    public void deleteUser(@RequestParam(name="userId") String userId){
-         userService.deleteUser(userId);
+    public void deleteUser(@RequestParam(name="userId") String userId, @RequestHeader(name="idToken") String idToken) throws IOException, FirebaseAuthException {
+         FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+         if(firebaseUser != null) {
+             userService.deleteUser(userId);
+         }
     }
